iac/pyproject.toml
[tool.poetry]
name = "lab"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.scripts]
labcli = "lab.__main__:app"

[tool.poetry.dependencies]
python = "^3.11"
cdktf = "^0.20.7"
typer = "^0.12.3"
cdk8s = "^2.68.77"
pyyaml = "^6.0.1"
pydantic = "^2.7.4"
cdk8s-plus-29 = "^2.3.10"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
ruff = "^0.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
